Add function to upload videos to s3
