+++ scripts/Pipfile
@@ -1,0 +1,33 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+boto3 = "==1.26.8"
+botocore = "==1.29.8"
+certifi = "==2022.9.24"
+charset-normalizer = "==2.1.1"
+colorama = "==0.4.6"
+decorator = "==4.4.2"
+idna = "==3.4"
+imageio = "==2.22.4"
+imageio-ffmpeg = "==0.4.7"
+jmespath = "==1.0.1"
+moviepy = "==1.0.3"
+numpy = "==1.23.4"
+pillow = "==9.3.0"
+proglog = "==0.1.10"
+python-dateutil = "==2.8.2"
+python-dotenv = "==0.21.0"
+pytube = "==12.1.0"
+requests = "==2.28.1"
+s3transfer = "==0.6.0"
+six = "==1.16.0"
+tqdm = "==4.64.1"
+urllib3 = "==1.26.12"
+
+[dev-packages]
+
+[requires]
+python_version = "3.9"
