+++ scripts/README.md
@@ -1,0 +1,18 @@
+### Install packages
+
+Using pipenv:
+```
+pipenv install
+```
+
+Using venv:
+```
+pip install -r requirements.txt
+```
+### Run the script:
+
+```
+python script.py <video_list.csv> <download_path> <extracted_ads_path> <s3_bucket_name>
+```
+
+Run `python script.py --help` for more details about the arguments.
