+++ scripts/env.sample
@@ -1,0 +1,2 @@
+AWS_ACCESS_KEY = ""
+AWS_SECRET_ACCESS_KEY = ""
